refactor(minimax): isolate remains API client
